+++ core-bundle/src/Resources/contao/classes/BackendUser.php
@@ -382,7 +382,7 @@ class BackendUser extends \User
 		\Config::set('useCE', $this->useCE);
 		\Config::set('thumbnails', $this->thumbnails);
 		\Config::set('backendTheme', $this->backendTheme);
-		\Config::set('limitWidth', $this->limitWidth);
+		\Config::set('fullscreen', $this->fullscreen);
 
 		// Inherit permissions
 		$always = array('alexf');